communicator: introduce OMPI_COMM_DISJOINT flag
This patch introduces a new communicator flag to indicate if no
processes share the same node.

This flag is intended for optimization of hierarchy-aware collective
operations to select the more efficient transport/algorithm.

In this change the flag is set during communicator activation via an
allreduce op, which incurs a one-time overhead for new communicators.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>