communicator: introduce OMPI_COMM_DISJOINT flag
This patch introduces a new communicator flag to indicate if no
processes share the same node.

This flag is intended for optimization of hierarchy-aware collective
operations to select the more efficient transport/algorithm.

In this change we introduce a non-blocking allreducestep in communicator
activation and set the new flag in the completion callback function.

Signed-off-by: Wenduo Wang <wenduwan@amazon.com>