mca/coll: add bruck method for allgather for any radix k
The allgather method is based on Bruck's original paper "Efficient
Algorithms for All-to-all Communications in Multiport Message-Passing
Systems" and a extension of current implementation in Open MPI
(ompi_coll_base_allgather_intra_bruck). The new method use a fanout of
k to allow multiple messages simultaneously send and receive on each
node to reduce the number of communication steps required to complete
the allgather operation (from log2(n) to logk(n)).

Signed-off-by: Jun Tang <juntangc@amazon.com>